rere

+++ src/main/java/com/beyond/order/ordering/dtos/OrderingCreateDto.java
@@ -21,7 +21,7 @@ public class OrderingCreateDto {
     @NotBlank
     private Long productId;
     @NotBlank
-    private int productCount;
+    private Long productCount;
     public static Ordering toEntity() {
         return Ordering.builder()
                 .orderStatus(OrderStatus.ordered)
